Fix rseq_registered() to compare rseq_size as signed
rseq_size is -1U until a successful rseq_init(), and rseq_registered()
compared it as unsigned: the sentinel read as ~4 billion, so an
uninitialized library answered "registered".  A caller that believes it
goes on to build an rseq area pointer from rseq_offset, which is still
PTRDIFF_MIN at that point, and dereferences it.

The three sibling predicates over rseq_size -- rseq_node_id_available(),
rseq_mm_cid_available() and rseq_slice_ctrl_available() -- already cast to
int for exactly this reason.  rseq_registered() was the only one that did
not.

This also covers the initialization-failed case: a libc that does not
export the rseq symbols makes rseq_init() return
RSEQ_INIT_ERROR_MISSING_SYMBOLS and leaves rseq_size at -1U, so the
predicate now answers false both before initialization and after a failed
one.

+++ include/rseq/rseq.h
@@ -138,11 +138,17 @@ bool rseq_available(unsigned int query);
 
 /*
  * Returns true if rseq is registered.
+ *
+ * Compared as signed, like the feature predicates below: rseq_size is -1U
+ * until a successful rseq_init(), and an unsigned comparison reads that
+ * sentinel as ~4 billion and answers "registered" for an uninitialized
+ * library -- while rseq_offset is still PTRDIFF_MIN, so a caller that
+ * believes it goes on to build an rseq area pointer from a garbage offset.
  */
 static inline __attribute__((always_inline))
 bool rseq_registered(void)
 {
-	return rseq_size > 0;
+	return (int) rseq_size > 0;
 }
 
 /*
